// src/app/models/text-prop.ts
export interface TextProp {
  id: number;
  text: string;
  textSize: string;
  textWeight: string;
  textStyle: string;
  subText: string;
  subTextSize: string;
  subTextWeight: string;
  subTextStyle: string;
  color: string;
  backgroundColor: string;
  toggleSide: boolean;
  toggleColor: string;
  toggleFlip: boolean;
  image: string;
  imageTintColor: string;
  imageDarken: string;
  imageToggle: string;
  alignText: string;
  justifyText: string;
}